Add ability to get delivery addresses

+++ src/back/Plants.Application/Requests/AddressRequest.cs
@@ -4,4 +4,4 @@ namespace Plants.Application.Requests;
 
 public record AddressRequest() : IRequest<AddressResult>;
 public record AddressResult(List<PersonAddress> Addresses);
-public record PersonAddress(string City, int MailNumber);
+public record PersonAddress(string City, long MailNumber);
